Phase 1: split Question & Quiz Analysis into two pages, nest SPV in the nav
The combined page is now pages/1_Question_Analysis.py (sections 1-7) and
pages/3_Quiz_Analysis.py (sections 8-13), sliced from the original file so every
moved block is byte-identical. Solution Process Visualization becomes
pages/2_Solution_Process_Visualization.py and renders indented, one font step
down, behind a turn-down arrow in the sidebar nav — presentational nesting only,
since Streamlit's page nav is a flat list with no parent-page concept.

build_question_analytics moves to analytics/question_analytics.py: the page files
are prefixed with digits to control nav order, which makes them non-importable,
and the Question Analysis page, the test suite and (next phase) the shared PDF
builder all need it.

Each page keeps a PDF panel scoped to its own sections for now; phase 5 replaces
both with one shared builder.
